requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.25.2
Pillow==10.1.0
